Bank_Administration_System/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
mysql-connector-python==8.1.0
bcrypt==4.0.1
python-dotenv==1.0.0
Flask-Login==0.6.3 